fix: cast total fees to integer and update national ID string labels

+++ app/src/main/java/com/example/yourdigitalpath/presentation/uploadfile/ServiceSummaryScreen.kt
@@ -287,7 +287,7 @@ fun ServiceSummaryScreen(
                                 Text(
                                     text = stringResource(
                                         R.string.total_price,
-                                        requestState.totalFees
+                                        requestState.totalFees.toInt()
                                     ),
                                     fontSize = 36.sp,
                                     fontWeight = FontWeight.Bold,
